Made some changes on the styling and Code structure
