+++ assets/style.css
@@ -77,18 +77,74 @@ h2 {
 /* Beginning of sign up main */
 
 
-/* h3 {
-    text-align: left; 
- } */
+h3 {
+    text-align: center;
+    font-size: 10px; 
 
-  .formheader{ 
-     padding: 12px;
+ } 
+
+  .form-wrapper-1{ 
+     width: 300px;
+     height: 320px;
+     margin: 0 auto;
+     position: relative;
+     
+ } 
+
+.form-div-1 {
+    display: block;
+     width: 100%;
+     height: 100%;
+     position: absolute;
      background-color: #EFEFEF; 
  } 
-  
+
+
+
+form > input {
+    display: inline-block;
+    justify-content: center;
+    align-items: center;
+    border: none;
+    border-radius: 3px;
+    padding-top: 10px;
+    padding-right: 20px;
+    margin-left: 70px;
+    margin-bottom: 10px;
+    font-weight: lighter;
+    
+}
+
+.button {
+    background-color: #000;
+    color: #FFFFFF;
+    font-weight: 200;
+    display: inline-block;
+    text-align: center !important;
+    margin-top: 10px;
+    padding: 10px 15px;
+    border: none;
+}
+
+.sidebar{
+    position: absolute;
+    top: 200px;
+    right: 0px;
+    width: 300px;
+   
+
+}
+
+.block{
+    width: 50%;
+	display: inline-block;
+	-webkit-box-sizing: border-box;
+	box-sizing: border-box;
+    -moz-box-sizing: border-box; 
+}
  .advert-images {
      width:200px;
-     height: 200px;
+     height: 100px;
      float: right;
      margin-bottom: 10px;
      position: relative;
@@ -98,14 +154,16 @@ h2 {
   /* h4 {
     float: right;
     padding-left: 10px;
-    bottom: 10px;     
- } */
+    bottom: 10px;  
+    font-size: 11px;
+   
+ }  */
 
 
 /* End of sign up main */
 
 /* Beginning of sign in main */
-.section-1 {
+.form-wrapper-2 {
     width: 300px;
     height: 200px;
     margin: auto;
@@ -113,6 +171,7 @@ h2 {
     background-color: #EFEFEF;
     position: relative;
 }
+
 
 .section-1 h4 {
     text-align: center;
@@ -122,22 +181,27 @@ h2 {
     text-align: center;
 }
 
-.form-div {
- 
-  width: 50px;
-  height: 50px;
-  /* Center vertically and horizontally */
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  margin: -25px 0 0 -25px;
+.form-div-2{
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    background-color: #EFEFEF; 
+} 
 
+.button-2 {
+    background-color: #000;
+    color: #FFFFFF;
+    font-weight: 200;
+    display: inline-block;
+    text-align: center !important;
+    border-radius: 3px;
+    margin-top: 10px;
+    margin-left: 70px;
+    padding: 10px 15px;
+    border: none;
 }
 
 /* End of signin main */
-
-
-
